birthday card listview refactor
